xml funciona, en proceso el dao completo

+++ ACD/NewspaperDB/src/main/java/ui/pantallas/newspaper/delete/NewsDeleteController.java
@@ -4,12 +4,18 @@ import domain.modelo.Newspaper;
 import jakarta.inject.Inject;
 import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
+import javafx.scene.control.Alert;
+import javafx.scene.control.ButtonType;
 import javafx.scene.control.TableColumn;
 import javafx.scene.control.TableView;
 import javafx.scene.control.cell.PropertyValueFactory;
+import javafx.stage.WindowEvent;
 import ui.pantallas.common.BasePantallaController;
 
 import java.io.IOException;
+import java.util.Optional;
+
+import static java.awt.SystemColor.window;
 
 public class NewsDeleteController extends BasePantallaController {
 
@@ -53,11 +59,29 @@ public class NewsDeleteController extends BasePantallaController {
     }
 
     @FXML
-    private void delete(ActionEvent actionEvent) {
+    private void delete() {
         if (tableNews.getSelectionModel().getSelectedItem() == null) {
-            getPrincipalController().sacarAlertError("error we");
+            getPrincipalController().sacarAlertError("Error, nothing selected.");
         } else {
-            newsDeleteViewModel.deleteNewspaper(tableNews.getSelectionModel().getSelectedItem().getNewspaperID());
+            if (newsDeleteViewModel.containsArticels(
+                    tableNews.getSelectionModel().getSelectedItem().getNewspaperID())){
+                Alert alertDelete = new Alert(Alert.AlertType.INFORMATION);
+                alertDelete.getButtonTypes().remove(ButtonType.OK);
+                alertDelete.getButtonTypes().add(ButtonType.CANCEL);
+                alertDelete.getButtonTypes().add(ButtonType.YES);
+                alertDelete.setTitle("Delete");
+                alertDelete.setContentText("This newspaper contains articles, delete anyway?");
+                Optional<ButtonType> res = alertDelete.showAndWait();
+
+
+                res.ifPresent(buttonType -> {
+                    if (buttonType == ButtonType.YES) {
+                        newsDeleteViewModel.deleteNewspaper(tableNews.getSelectionModel().getSelectedItem().getNewspaperID());
+                    }
+                });
+            } else {
+                newsDeleteViewModel.deleteNewspaper(tableNews.getSelectionModel().getSelectedItem().getNewspaperID());
+            }
         }
     }
 }
